+++ modules/apps/layout/layout-page-template-service/src/main/java/com/liferay/layout/page/template/internal/headless/delivery/dto/v1_0/structure/LayoutStructureItemExportHelperTrackerImpl.java
@@ -38,8 +38,8 @@ public class LayoutStructureItemExportHelperTrackerImpl
 	implements LayoutStructureItemExportHelperTracker {
 
 	@Override
-	public LayoutStructureItemExportHelper
-		getLayoutStructureItemExportHelper(String className) {
+	public LayoutStructureItemExportHelper getLayoutStructureItemExportHelper(
+		String className) {
 
 		List<LayoutStructureItemExportHelper>
 			layoutStructureItemExportHelperList = new ArrayList<>(
